Default layout for versioned services
