+++ app/models/portfolio.rb
@@ -1,7 +1,7 @@
 class Portfolio < ApplicationRecord
   include OwnerField
   include Discard::Model
-  include Catalog::DiscardRestore
+  include Api::V1x0::Catalog::DiscardRestore
   include Aceable
   include UserCapabilities
