+++ app/models/portfolio_item.rb
@@ -1,8 +1,8 @@
 class PortfolioItem < ApplicationRecord
   include OwnerField
   include Discard::Model
-  include Catalog::DiscardRestore
   include UserCapabilities
+  include Api::V1x0::Catalog::DiscardRestore
   destroy_dependencies :service_plans
 
   acts_as_tenant(:tenant)
