+++ app/services/api/v1x0/catalog.rb
@@ -1,0 +1,9 @@
+#require 'catalog/service_offering'
+require 'api/v1x0/catalog/cancel_order'
+module Api
+  module V1x0
+    module Catalog
+
+    end
+  end
+end
