+++ app/services/api/v1x0/catalog/create_request_for_applied_inventories.rb
