+++ app/services/api/v1x0/catalog/next_name.rb
