+++ app/services/api/v1x0/catalog/order_item_sanitized_parameters.rb
