+++ app/services/api/v1x0/catalog/share_resource.rb
