+++ app/services/api/v1x0/catalog/soft_delete_restore.rb
