+++ app/services/api/v1x0/catalog/update_icon.rb
