+++ app/services/api/v1x0/service_offering.rb
@@ -1,0 +1,7 @@
+module Api
+  module V1x0
+    module ServiceOffering
+
+    end
+  end
+end
