+++ app/services/api/v1x0/tags.rb
@@ -1,0 +1,7 @@
+module Api
+  module V1x0
+    module Tags
+
+    end
+  end
+end
