+++ app/services/api/v1x1/group.rb
@@ -1,0 +1,7 @@
+module Api
+  module V1x0
+    module Group
+
+    end
+  end
+end
